+++ src/store/modalStore.ts
@@ -1,0 +1,22 @@
+import { create } from 'zustand';
+
+type ModalType = 'alert' | 'card' | null;
+
+interface ModalState {
+  isOpen: boolean;
+  modalType: ModalType;
+  openModal: (type: ModalType) => void;
+  closeModal: () => void;
+}
+
+export const modalStore = create<ModalState>((set) => ({
+  isOpen: false,
+  modalType: null,
+  openModal: (type) => {
+    set({ isOpen: true, modalType: type });
+  },
+
+  closeModal: () => {
+    set({ isOpen: false, modalType: null });
+  },
+}));
